requirements.txt
beautifulsoup4
pyodbc
sqlalchemy
pandas
requests
python-dotenv
python-decouple
streamlit
Pillow
xlsxwriter